issue-88-upgrade: Fixing issue related to 32-bit builds, adding random number for flag in testPrepareWithParam() unit test.

+++ native/sqlite3_wrap_manual.c
@@ -887,7 +887,7 @@ JNIEXPORT jint JNICALL Java_com_almworks_sqlite4java__1SQLiteManualJNI_sqlite3_1
 
 
 JNIEXPORT jint JNICALL Java_com_almworks_sqlite4java__1SQLiteManualJNI_sqlite3_1win32_1set_1directory(JNIEnv *jenv, jclass jcls,
-  jlong jtype, jstring zValue)
+  jint jtype, jstring zValue)
 {
 #ifdef SQLITE_OS_WIN
   int rc;
